requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
msgpack==1.0.8
pynvim==0.5.0
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.2
